Print the size of a term on each reduction step
It seems that there were actually no "memory leaks"; it turned out that
a number of sub-terms sometimes skyrockets to hundreds of millions or
even tens of billions. No wonder the program eats up the entire memory.

Though not proved/confirmed, I believe the cause is substitutions of a
large sub-term into a lot of places.

Solving the issue is not easy; I would have to implement CSE or
something. As a quick fix, I'm going to put a limit of sub-terms in a
term.