Deep-reduce in reduceStep to properly handle infinite loops
It seems reducing some lambda terms reproduces nested sub-terms
identical to itself. Improper recursive function got into an infinite
loop because of it, and it looks like that it is this that the root
cause of the program hang.

Such a "pathological(?, I'm not yet to see a concrete example)"
reduction happens when deep-reducing a second term of App. It is now
done in 'reduceStep', and @C.take 1000@ now takes care of it.