Fix interpretChurchNumber to avoid conflict of free variables
@interpretChurchNumber (Abs (Abs (Var 3)))@ incorrectly returned
@Just 0@ because applying two variables reduces it to @Var 1@, which
coincides with the Church numeral 0.

This is a bug because an open term containing a (3rd) free variable is
not a Church numeral.

@interpretChurchNumber@ now avoids conflicts of free variables by first
determining the largest free variable referenced in the term, and
applies two variables larger than that.

Now the unit test for _churchNumber reports that it is a prism.